Классы третьей части перенесены в соответствующий пакет и частично отрефакторены.
